Read collectors via scorecardCollectorsServiceRef
