+++ workspaces/scorecard/plugins/scorecard-backend/docs/providers.md
@@ -4,8 +4,8 @@ The Scorecard plugin collects metrics from third-party data sources using metric
 
 For provider data-fetching reuse across datasources, Scorecard provides collector contracts:
 
-- `scorecardCollectorsExtensionPoint` to register collectors
-- `collectWithContract(...)` helper to call a collector with dual validation:
+- `scorecardCollectorsExtensionPoint` extension point to register collectors
+- `scorecardCollectorsServiceRef` service to call `collect(...)` and read collector output with dual validation:
   - provider expected input/output schemas
   - collector declared input/output schemas
